requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
Pillow>=12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
django-editorjs-fields==0.2.7